Rename Emitter.ts -> Channel.ts
