+++ src/Channel.ts
