+++ src/index.ts
@@ -1,4 +1,4 @@
 export { on, once, put, take, takeEvery } from "./effects";
-export { Channel as Emitter } from "./Emitter";
+export { Channel } from "./Channel";
 export { Store } from "./Store";
 export { IAction, IActionType, IChannel, IStore } from "./types";
